feat(server): generate the REST using the OpenAPI spec
With this change, the REST server is now generated based on the OpenAPI
spec. This ties up both the definition and the implementation,
preventing further deviations and invalid specifications (if it's
invalid, it will fail to generate or won't be able to be used inside the
server).

With this, we can ensure that the OpenAPI consumers will always have the
correct definition